Struct succint_dbg_rs::SDbg

pub struct SDbg { /* fields omitted */ }

struct to implement succint dbg with

kmersize: the kmer size

n_nodes: amount of nodes

last: last array

edge: edges array

fvec: F array

neg: array to recognize edges marked as negative

rschar: map to obtain rank/select of edges

rscharneg: map to obtain rank/select of edges marked as negative

rslast: map to obtain rank/select of last array

Implementations

impl SDbg

pub fn new(reads: &mut Vec<String>, k: u32) -> Self

create new succint dbg from a set of strings

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
assert_eq!(sdbg.n_nodes(), 11)

pub fn new_from_string(read: &String, k: u32) -> Self

create new succint dbg from a string

Example

use succint_dbg_rs::SDbg;
let sdbg = SDbg::new_from_string(&"TACGACGTCGACT".to_string(), 4);
assert_eq!(sdbg.n_nodes(), 11)

pub fn kmersize(&self) -> u32

return the chosen k for kmers

pub fn n_nodes(&self) -> usize

return the amount of nodes in the succint dbg

pub fn last(&self) -> &Vec<usize>

return the last array (as explained in Bowe’s research)

pub fn edge(&self) -> &Vec<char>

return the vec of edges (as explained in Bowe’s research)

pub fn neg(&self) -> &Vec<bool>

return the bool array that it specifies if edge in marked as negative

pub fn fvec(&self) -> &Vec<usize>

return the F array (as explained in Bowe’s research)

pub fn rschar(&self) -> &HashMap<char, RankSelect>

return the map with the bitvec of a char, char used by key, with rank/select (as explained in Bowe’s research)

pub fn rscharneg(&self) -> &HashMap<char, RankSelect>

return the map with the bitvec of a char marked as neg, char used by key, with rank/select (as explained in Bowe’s research)

pub fn rslast(&self) -> &RankSelect

return the map with the bitvec of the last array with rank/select (as explained in Bowe’s research)

pub fn print(&self)

print main structires of succint dbg

pub fn findsymbol(&self, i: isize) -> char

return the last char at index F array

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut symb = Vec::new();
let symbcheck = vec!['$', 'A', 'A', 'C', 'C', 'C', 'C', 'G', 'G', 'G', 'T', 'T', 'T'];
for i in 0..sdbg.edge().len() {
    symb.push(sdbg.findsymbol(i as isize));
}
assert_eq!(symb, symbcheck);

pub fn first_edge(&self, i: isize) -> isize

return first edge from a node

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut edged = Vec::new();
let edgecheck = vec![0, 1, 2, 3, 5, 6, 7, 9, 10, 11, 12];
for i in 0..sdbg.n_nodes() {
    edged.push(sdbg.first_edge(i as isize));
}
assert_eq!(edged, edgecheck);

pub fn last_edge(&self, i: isize) -> isize

return last edge from a node

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut edged = Vec::new();
let edgecheck = vec![0, 1, 2, 4, 5, 6, 8, 9, 10, 11, 12];
for i in 0..sdbg.n_nodes() {
edged.push(sdbg.last_edge(i as isize));
}
assert_eq!(edged, edgecheck);

pub fn node_range(&self, i: isize) -> (isize, isize)

return range of edges from a node

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut ranged = Vec::new();
let rangecheck = vec![(0, 0), (1, 1), (2, 2), (3, 4), (5, 5), (6, 6), (7, 8),
                      (9, 9), (10, 10), (11, 11), (12, 12)];
for i in 0..sdbg.n_nodes() {
    ranged.push(sdbg.node_range(i as isize));
}
assert_eq!(ranged, rangecheck);

pub fn edge_to_node(&self, i: isize) -> isize

change edge index in node index

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut edged = Vec::new();
let edgecheck = vec![0, 1, 2, 3, 3, 4, 5, 6, 6, 7, 8, 9, 10];
for i in 0..sdbg.edge().len() {
    edged.push(sdbg.edge_to_node(i as isize));
}
assert_eq!(edged, edgecheck);

pub fn forward(&self, i: isize) -> isize

return index of the last edge of the node pointed to by edge as parameter (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut forw = Vec::new();
let forcheck = vec![10, 4, 5, 8, 11, 8, 9, 1, 12, 1, 2, -1, 6];
for i in 0..sdbg.edge().len() {
    forw.push(sdbg.forward(i as isize));
}
assert_eq!(forw, forcheck);

pub fn backward(&self, i: isize) -> isize

return index of the first edge that points to the node that the edge at index exits (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut backw = Vec::new();
let backcheck = vec![-1, 7, 10, 1, 1, 2, 12, 3, 3, 6, 0, 4, 8];
for i in 0..sdbg.edge().len() {
    backw.push(sdbg.backward(i as isize));
}
assert_eq!(backw, backcheck);

pub fn outdegree(&self, i: isize) -> isize

return number of outgoing edges from a node (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut edged = Vec::new();
let edgecheck = vec![1, 1, 1, 2, 1, 1, 2, 1, 1, 1, 1];
for i in 0..sdbg.n_nodes() {
    edged.push(sdbg.outdegree(i as isize));
}
assert_eq!(edged, edgecheck);

pub fn indegree(&self, i: isize) -> isize

return number of incoming edges to a node (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut ind = Vec::new();
let incheck = vec![0, 2, 1, 1, 1, 1, 2, 1, 1, 1, 1];
for i in 0..sdbg.n_nodes() {
    ind.push(sdbg.indegree(i as isize));
}
assert_eq!(ind, incheck);

pub fn outgoing(&self, i: isize, symbol: char) -> isize

return index obtained from a node, follow the edge labeled by a symbol (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut edgeg = Vec::new();
let edgecheck = vec![-1, -1, -1, -1, 8, -1, -1, 3, -1, -1, -1, -1, 4, -1, -1, -1, -1, -1,
                      6, 9, -1, -1, -1, 6, -1, -1, -1, -1, 7, -1, -1, 1, -1, -1, 10, -1, 1,
                      -1, -1, -1, -1, 2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 5, -1, -1];
for i in 0..sdbg.n_nodes() {
    for j in ['$', 'A', 'C', 'G', 'T'] {
        edgeg.push(sdbg.outgoing(i as isize, j));
    }
}
assert_eq!(edgeg, edgecheck);

pub fn successors(&self, i: isize) -> Vec<isize>

return vec with index of successors from a node (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut edges = Vec::new();
let succcheck = vec![vec![8], vec![3], vec![4], vec![6, 9], vec![6], vec![7],
                     vec![1, 10], vec![1], vec![2], vec![-1], vec![5]];
for i in 0..sdbg.n_nodes() {
    edges.push(sdbg.successors(i as isize));
}
assert_eq!(edges, succcheck);

pub fn label(&self, i: isize) -> String

return label of a node (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut label = Vec::new();
let labelcheck = vec!["$$$", "CGA", "$TA", "GAC", "TAC", "GTC", "ACG", "TCG",
                      "$$T", "ACT", "CGT", "$$T", "CGA"];
for i in 0..sdbg.edge().len() {
    label.push(sdbg.label(i as isize));
}
assert_eq!(label, labelcheck);

pub fn incoming(&self, i: isize, symbolf: char) -> isize

return predecessor node starting with a symbol, that has an edge to a node (as explained in Bowe’s research)

Example

use succint_dbg_rs::SDbg;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
let mut ing = Vec::new();
let incheck = vec![-1, -1, -1, -1, -1, -1, 6, -1, -1, 7, 8, -1, -1, -1, -1, -1, -1,
                   1, -1, -1, 2, -1, -1, -1, -1, -1, -1, 10, -1, -1, -1, -1, -1, 3, 4, -1,
                   -1, -1, 5, -1, 0, -1, -1, -1, -1, -1, -1, -1, 3, -1, -1, 6, -1, -1, -1];
for i in 0..sdbg.n_nodes() {
    for j in ['$', 'A', 'C', 'G', 'T'] {
        ing.push(sdbg.incoming(i as isize, j));
    }
}
assert_eq!(ing, incheck);

pub fn to_dot(&self, edgeput: &str)

print to file of the .dot of tne graph

Example

use succint_dbg_rs::SDbg;
use std::path::Path;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
sdbg.to_dot("output/bowe.dot");
assert!(Path::new("output/bowe.dot").exists());

pub fn to_dot_no_dollar(&self, edgeput: &str)

print to file of the .dot of tne graph without $ nodes

Example

use succint_dbg_rs::SDbg;
use std::path::Path;
let mut kmers = vec!["TACGACGTCGACT".to_string()];
let sdbg = SDbg::new(&mut kmers, 4);
sdbg.to_dot_no_dollar("output/bowend.dot");
assert!(Path::new("output/bowend.dot").exists());